scave/python: Fix missing file extension on image/data exported from charts
When the provided export filename (or the chart name, as a fallback)
"looked like" it already had a 3 or 4 character file extension at the end
(eg. "Delay in client[3].app"), the extension for the selected file format
was not appended to the file name. This resulted in either the data or the
image file overwriting the other, without an appropriate file extension.