qtenv: Added simulation identicon and "Show Simulation Info" functionality
The identicon appears in toolbar, and its purpose is to help the user
tell apart Qtenv windows when several similar simulations are run
concurrently (e.g. for finding the source of regressions.)

Identicon is drawn using a "simulation hash" as seed, which includes
omnetpp version, the config name and run number, the inifile name,
the working directory, the configuration contents, the RNGs,
module names, their NED and C++ type names, their parameter names and
values, etc.

"Show Simulation Info" displays a report of all values that affect the
simulation hash, and much more.