ide/scave: Make selected rows in DataTables legible on macOS in dark mode
Selected rows were white text on white background,
due to (at least) SWT.COLOR_LIST_SELECTION returning
a color with some transparency, and the GC that used
it to paint the selected rows not taking it into
account (it has a separate setAlpha() method).

Instead of painting the background of selected
rows with transparency every time, let's resolve
the final color beforehand.