ide/ned: fix #1343 "NED Inheritance View usage inconvenience"
Issue: The NED Inheritance View only shows content when the cursor is positioned
INSIDE the component definition in the NED file. This is especially annoying
when the NED file only contains one component definition. But even when the
file contains multiple ones, it could display info e.g. related to the one
closes to the cursor.

Solution: Return the module (or any other type) that follows the cursor
as selection (unless the cursor is positioned into something specific in the
file)