ide/ned: fix #1341 "ConcurrentModificationException error on saving NED file."
The ConcurrentModificationException that was occurring when saving
NED files. The issue was in the NedTypeInfo class, where collections
were being modified while they were being iterated over.

The problem occurred in the refreshLocalMembersIfNeeded() method, which was
modifying collections that might be in use by other parts of the code.
The specific outer loop causing the issue was in
refreshInheritedMembersIfNeeded:

for (INedTypeInfo typeInfo : ancestors) {
    Assert.isTrue(typeInfo instanceof NedTypeInfo);
    NedTypeInfo component = (NedTypeInfo)typeInfo;

Solution: Do not modify collections that were returned to the user.
Instead, create new ones and changes the references in NedTypeInfo
point to the new ones. For added measure, also make those collections
unmodifiable.