ide/ini: Form-based editor: fix wrong foreground colors (#1356 issue 2)
The theme engine apparently runs asynchronously, and overwrites the fg color
set in the field editor constructor. Moreover, setForeground(null) does
not work as expected, also due to interference with the theme engine.

Solution is not to fight the theme engine (it is a losing battle),
but instead go with it and do the graying using CSS.